+++ ZoomLevel/ModConfig.cs
@@ -34,5 +34,7 @@ namespace ZoomLevel
         public bool SuppressControllerButton { get; set; } = true;
         public bool ZoomAndUIControlEverywhere { get; set; } = false;
         public bool IsHideUIWithCertainZoom { get; set; } = false;
+
+        public bool PressAnyButtonToCenterCamera { get; set; } = true;
     }
 }